made print list function, tested functions and add catch error logic
In the function of printing the list, I just print the values separated by n, but I check if the past list points to 0(null), if it happens, I deal with an error function, which prints the message and closes the program. Also, I have tested the print and insert functions in the list and they are working perfectly.