Change list_print to print history of results
Considering that this function should print all the elements of the list except the first one which would be the first entry, it makes no sense for this to be a function of the list, because it is not generic. Thus, a function was created that empties the list by printing the values, because if it is not a function of the list, its nodes or structure should not be manipulated.